✨ Affichage MAP

+++ src/Sprites/Sprite.cpp
@@ -8,11 +8,11 @@ namespace Sprites
 
 Sprite::Sprite() : Sprite{0,0,0} {}
 
-Sprite::Sprite(const std::array<int, 3> &pos) : Sprite{pos, 1, 0} {}
+Sprite::Sprite(const std::array<float, 3> &pos) : Sprite{pos, 1, 0} {}
 
-Sprite::Sprite(int x, int y, int zIndex) : Sprite{{x,y,zIndex}, 1, 0}{}
+Sprite::Sprite(float x, float y, float zIndex) : Sprite{{x,y,zIndex}, 1, 0}{}
 
-Sprite::Sprite(const std::array<int, 3> &position, float scale, float rotation) :
+Sprite::Sprite(const std::array<float, 3> &position, float scale, float rotation) :
 position_{position[0], position[1]}, zindex{position[2]}, scale_{scale}, rotation_{rotation}
 {
